uidai_analytics/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
celery==5.3.4
django-celery-results==2.5.1
redis==5.0.1
django-redis==5.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
django-cors-headers==4.3.1
pandas==2.1.4
scikit-learn==1.3.2
prophet==1.1.5
pmdarima==2.0.4
joblib==1.3.2
sqlalchemy==2.0.25
PyJWT==2.8.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10